Simplify namespace for files
