+++ ruby/lib/cucumber/.gitignore
@@ -1,0 +1,1 @@
+ci_environment/CiEnvironments.json
