--- ruby/lib/cucumber/ci_environment/.gitignore
@@ -1,1 +1,0 @@
-CiEnvironments.json
